# docs/koready-backend-design/openapi.yaml
openapi: 3.0.3
info:
  title: Koready API
  version: 1.0.0
  description: |
    Koready MVP 사용자 API와 OpenAPI 공모전 증빙 관리자 API의 기준 계약입니다.

    각 operation의 설명은 호출 화면, 선행 조건, 성공 후 프론트 행동을 포함합니다.
    `x-implementation-status=PLANNED`인 API는 Swagger 예제와 mock에만 사용하고 테스트 서버에서 호출하지 않습니다.
    전체 화면 흐름과 오류 처리 예제는 `10_FRONTEND_API_FLOW_GUIDE.md`를 참고합니다.
    로그인 사용자의 다국어 응답은 프로필에 저장한 KO/EN 설정을 우선합니다.
    비로그인 공개 조회 또는 아직 사용자 레코드가 없는 테스트 요청만 Accept-Language를 사용합니다.
    장소 카드와 상세의 saved/isSaved는 로그인 사용자의 실제 저장 상태이며 비로그인 조회에서는 false입니다.
servers:
  - url: /api/v1
...
security:
  - bearerAuth: []
...
paths:
...
  /recommendation-decks:
    post:
      tags: [Recommendation]
      operationId: createRecommendationDeck
      summary: K-Local Pick 추천 덱 생성
      description: |
        **호출 시점:** 추천 탭에 처음 들어가거나 근교/전국 범위를 새로 선택할 때 호출합니다.
        NEARBY는 실제 거리 반경이 아니라 출발 위치와 같은 서비스 권역입니다. originLocationId를 생략하면 기본 위치를 사용합니다.
        응답의 카드 순서를 그대로 표시하고 deckId와 nextCursor를 다음 페이지 호출까지 보관합니다.
        현재 데이터 단계에서는 사용자 여행 스타일이 장소 스타일과 일치하면 matchRank=2,
        일치하지 않는 확장 후보는 matchRank=3입니다. preference tag가 연결되는 온보딩 구현 뒤
        태그와 유형이 모두 일치하는 matchRank=1을 활성화합니다.
        deckId와 cursor는 내용을 해석할 수 없는 서버 저장 무작위 값이며 덱은 생성 뒤 24시간 동안 조회할 수 있습니다.
      x-implementation-status: IMPLEMENTED
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RecommendationDeckRequest'
      responses:
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '201':
          $ref: '#/components/responses/RecommendationDeckCreated'
        '422':
          $ref: '#/components/responses/UnprocessableEntity'
  /recommendation-decks/{deckId}:
    get:
      tags: [Recommendation]
      operationId: getRecommendationDeckPage
      summary: 추천 덱의 다음 카드 조회
      description: |
        **호출 시점:** 현재 카드 묶음이 소진되고 hasMore=true일 때 nextCursor로 호출합니다.
        같은 deckId와 cursor 재요청은 같은 순서와 결과를 반환합니다. 프론트는 응답을 다시 섞거나 중복 제거 규칙을 재구현하지 않습니다.
        cursor를 생략하면 첫 페이지를 다시 조회하며 CARD_SERVED 기록과 30일 제한은 늘어나지 않습니다.
        페이지를 처음 반환하기 직전에 사용자 전체 노출 기록을 다시 확인합니다. 다른 덱에서 먼저 노출된 장소는
        이 페이지에서 제외되므로 cards가 요청한 size보다 적거나 비어 있을 수 있으며, 같은 cursor 재요청에는
        처음 확정된 결과를 그대로 반환합니다.
        다른 사용자의 덱, 만료된 덱과 잘못된 cursor는 모두 404로 처리해 덱 존재 여부를 노출하지 않습니다.
      x-implementation-status: IMPLEMENTED
      parameters:
        - $ref: '#/components/parameters/DeckId'
        - $ref: '#/components/parameters/Cursor'
      responses:
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
        '200':
          $ref: '#/components/responses/RecommendationDeckSuccess'
  /recommendation-decks/{deckId}/events:
    post:
      tags: [Recommendation]
      operationId: createRecommendationEvent
      summary: 추천 카드 탐색 행동 기록
      description: |
        카드 펼침, 이전·다음, 상세 진입, 저장, 경로 진입을 기록하는 분석용 API입니다.
        프론트는 사용자 동작마다 한 번만 비차단으로 전송하고 실패해도 자동 재시도하지 않습니다.
        이 API 실패 때문에 카드 이동, 상세 진입, 저장 상태 등 화면 핵심 동작을 되돌리거나 막지 않습니다.
        PLACE_SAVED/UNSAVED는 실제 저장 API 성공 뒤 전송합니다.
        CARD_SERVED는 서버가 응답할 때 기록하므로 프론트에서 보내지 않습니다.
        occurredAt은 분석용 사용자 동작 시각이며, 생략하면 서버 수신 시각을 사용합니다. 이 값은 30일 재노출 제한의
        시작 시각이나 기간을 바꾸지 않습니다. 응답의 recordedAt은 서버 DB 기록 시각입니다.
        로그인 사용자가 소유한 덱에서 서버가 이미 노출된 카드만 기록할 수 있습니다. 없는 덱, 다른 사용자 덱,
        덱에 없거나 아직 노출되지 않은 카드는 존재 여부를 구분하지 않고 모두 404로 응답합니다.
      x-implementation-status: IMPLEMENTED
      parameters:
        - $ref: '#/components/parameters/DeckId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RecommendationEventRequest'
      responses:
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
        '201':
          $ref: '#/components/responses/RecommendationEventCreated'
...
  /places/{placeId}:
    get:
      tags: [Place]
      operationId: getPlace
      summary: 관광지 상세 조회
      description: |
        **호출 시점:** 월별 추천, 지도, 검색, 추천 덱, 저장 목록의 카드를 눌렀을 때 호출합니다.
        제목, 주소, 이미지, nullable 운영정보, 태그, 설명, 관련 장소, 저장 상태, availableTabs를 반환합니다.
        프론트는 availableTabs에 있는 DESCRIPTION, ROUTE, MATES만 노출하고 null 운영정보 행은 숨깁니다.
        로그인 전에도 호출할 수 있고 isSaved는 false입니다.

        **현재 MVP 데이터 범위:** KTO/관리 데이터에 근거가 있는 제목, 권역, 주소, 좌표, 대표 이미지와
        설명만 반환합니다. 아직 수집하지 않은 운영시간·휴무일·요금·주차 정보는 추측하지 않고 null입니다.
        대표 이미지가 없으면 images는 빈 배열이며, 설명 원문이 없으면 description은 null이고
        availableTabs에 DESCRIPTION이 포함되지 않습니다. Buddy 메이트 조회가 구현되어 MATES는
        모든 표출 장소에 포함합니다. TMAP Route는 아직 PLANNED이므로 ROUTE는 포함하지 않습니다.

        **언어:** MVP에서는 `KTO_EN` 또는 운영자 수정 영문 제목이 있는 장소만 조회됩니다.
        Accept-Language가 en 계열이고 영문 소개가 없으면 overview는 null이며 한국어 또는 AI 소개로 대체하지 않습니다.
      security: []
      x-implementation-status: IMPLEMENTED
      parameters:
        - $ref: '#/components/parameters/PlaceId'
      responses:
        '400':
          $ref: '#/components/responses/BadRequest'
        '200':
          description: 장소 상세
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PlaceDetailEnvelope'
        '404':
          $ref: '#/components/responses/NotFound'
  /users/me/saved-places:
    get:
      tags: [Place]
      operationId: getMySavedPlaces
      summary: 내가 저장한 관광지 목록 조회
      description: |
        **호출 시점:** 하단 저장 탭에 들어갈 때 호출합니다.
        cursor 무한 스크롤을 사용하며 저장 취소 성공 시 해당 item을 로컬 목록에서 제거합니다.
        저장 이후 공개 조건을 잃은 장소는 목록에서 숨기며, 영문 소개가 없으면 overview는 null입니다.
      x-implementation-status: IMPLEMENTED
      parameters:
        - $ref: '#/components/parameters/Cursor'
        - $ref: '#/components/parameters/Size50'
      responses:
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '200':
          $ref: '#/components/responses/SavedPlaceListSuccess'
  /users/me/saved-places/{placeId}:
    put:
      tags: [Place]
      operationId: savePlace
      summary: 관광지 저장
      description: |
        **호출 시점:** 카드 또는 상세의 빈 하트를 누를 때 호출합니다.
        source에는 하트를 누른 화면을 전달합니다. 멱등 API이므로 이미 저장된 장소를 다시
        저장해도 최초 savedAt과 source를 유지하며 성공합니다. 저장 취소 후 다시 저장하면
        새 savedAt과 source로 복원됩니다. 성공 뒤 saved를 true로 갱신합니다.
      x-implementation-status: IMPLEMENTED
      parameters:
        - $ref: '#/components/parameters/PlaceId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SavePlaceRequest'
            examples:
              recommendationCard:
                summary: K-Local Pick 카드에서 저장
                value:
                  source: RECOMMENDATION_CARD
      responses:
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '200':
          $ref: '#/components/responses/SavePlaceSuccess'
        '404':
          $ref: '#/components/responses/NotFound'
    delete:
      tags: [Place]
      operationId: unsavePlace
      summary: 관광지 저장 취소
      description: |
        **호출 시점:** 활성 하트를 누를 때 호출합니다.
        204 body는 없습니다. 성공 뒤 isSaved를 false로 바꾸고 저장 목록 화면에서는 item을 제거합니다.
      x-implementation-status: IMPLEMENTED
      parameters:
        - $ref: '#/components/parameters/PlaceId'
      responses:
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '204':
          description: 저장 취소 성공
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: |
        staging과 prod에서는 실제 access token만 사용합니다.
        local 프로필의 Swagger에서는 실제 로그인 구현 전 공개 개발값인 local-user,
        local-operator, local-auditor, local-admin 중 하나를 입력할 수 있습니다.
        이 네 값은 local 전용이며 다른 프로필에서는 모두 401입니다.
  parameters:
...
    Cursor:
      name: cursor
      in: query
      description: 서버가 발급한 불투명 페이지 위치. 필터·정렬·언어가 같을 때만 그대로 재사용
      schema:
        type: string
        maxLength: 512
    Size50:
      name: size
      in: query
      schema:
        type: integer
        minimum: 1
        maximum: 50
        default: 20
...
    PlaceId:
      name: placeId
      in: path
      required: true
      schema:
        type: integer
        format: int64
...
    DeckId:
      name: deckId
      in: path
      required: true
      schema:
        type: string
...
  responses:
...
    RecommendationDeckCreated:
      description: 새 추천 덱 생성 성공
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/RecommendationDeckEnvelope'
    RecommendationDeckSuccess:
      description: 추천 덱의 다음 카드 묶음
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/RecommendationDeckEnvelope'
    RecommendationEventCreated:
      description: 추천 행동 기록 성공
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/RecommendationEventEnvelope'
...
    SavedPlaceListSuccess:
      description: 내가 저장한 관광지 cursor 목록
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/SavedPlaceListEnvelope'
    SavePlaceSuccess:
      description: 관광지 저장 상태
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/SavePlaceEnvelope'
...
    BadRequest:
      description: 입력 형식 오류
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ApiError'
    Unauthorized:
      description: 인증 실패
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ApiError'
...
    NotFound:
      description: 리소스 없음
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ApiError'
...
    UnprocessableEntity:
      description: 비즈니스 규칙 위반
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ApiError'
...
  schemas:
    ApiEnvelope:
      type: object
      required: [success, code, message, data, traceId]
      description: 공통 성공 메타데이터. 실제 data 타입은 각 operation 전용 Envelope가 정의함
      properties:
        success:
          type: boolean
          enum: [true]
        code:
          type: string
          example: OK
        message:
          type: string
        data:
          description: operation 전용 Envelope에서 구체 타입을 정의하는 성공 응답 본문
        traceId:
          type: string
    ApiError:
      type: object
      required: [success, code, message, traceId]
      properties:
        success:
          type: boolean
          enum: [false]
        code:
          type: string
        message:
          type: string
        errors:
          type: array
          description: 입력 필드별 오류. 비필드 오류에서는 빈 배열 또는 생략 가능
          items:
            $ref: '#/components/schemas/ApiFieldError'
        traceId:
          type: string
...
    SavedPlaceSource:
      type: string
      description: 사용자가 저장 버튼을 누른 화면. 추천 품질과 화면 사용성 분석에 사용
      enum: [HOME_MONTHLY, RECOMMENDATION_CARD, PLACE_DETAIL, MAP]
...
    TravelStyle:
      type: string
      description: 로컬 맛집, 지역 축제, 전통시장, 문화체험, 자연 명소, 전시·미술관, 드라마 촬영지
      enum: [LOCAL_FOOD, LOCAL_FESTIVAL, TRADITIONAL_MARKET, CULTURE_EXPERIENCE, NATURE, EXHIBITION_MUSEUM, DRAMA_LOCATION]
...
    ServiceRegionCode:
      type: string
      description: 서울, 경기·인천, 강원, 충청, 전라, 경상, 제주 7개 서비스 권역
      enum: [SEOUL, GYEONGGI, GANGWON, CHUNGCHEONG, JEOLLA, GYEONGSANG, JEJU]
...
    FestivalOccurrenceStatus:
      type: string
      description: Asia/Seoul 조회일과 해당 회차의 개최 기간으로 계산하는 축제 상태
      enum: [UPCOMING, ONGOING, ENDED]
    RecommendationScope:
      type: string
      description: NEARBY는 GPS 반경이 아니라 기본 위치와 같은 서비스 권역, NATIONWIDE는 전국
      enum: [NEARBY, NATIONWIDE]
    RecommendationEventType:
      type: string
      description: 추천 카드 탐색 이벤트. CARD_SERVED는 서버 전용이라 포함하지 않음
      enum: [CARD_EXPANDED, CARD_PREVIOUS, CARD_NEXT, PLACE_DETAIL_CLICKED, PLACE_SAVED, PLACE_UNSAVED, ROUTE_OPENED]
...
    SavePlaceRequest:
      type: object
      required: [source]
      properties:
        source:
          $ref: '#/components/schemas/SavedPlaceSource'
...
    RecommendationDeckRequest:
      type: object
      required: [scope, size]
      properties:
        scope:
          $ref: '#/components/schemas/RecommendationScope'
        originLocationId:
          type: integer
          format: int64
          nullable: true
        size:
          type: integer
          minimum: 1
          maximum: 50
    RecommendationEventRequest:
      type: object
      required: [placeId, eventType]
      properties:
        placeId:
          type: integer
          format: int64
          minimum: 1
          description: 현재 덱 응답에서 실제로 받은 카드의 장소 식별자
        eventType:
          $ref: '#/components/schemas/RecommendationEventType'
        occurredAt:
          type: string
          format: date-time
          nullable: true
          description: 분석용 사용자 동작 시각. 생략하면 서버 수신 시각을 사용하며 30일 재노출 제한에는 영향을 주지 않음
    PlaceDetailEnvelope:
      type: object
      required: [success, code, message, data, traceId]
      properties:
        success:
          type: boolean
          enum: [true]
        code:
          type: string
        message:
          type: string
        data:
          $ref: '#/components/schemas/PlaceDetail'
        traceId:
          type: string
    PlaceDetail:
      type: object
      required: [placeId, title, serviceRegionCode, locationText, address, latitude, longitude, images, tags, isSaved, description, relatedPlaces, availableTabs]
      properties:
        placeId:
          type: integer
          format: int64
          example: 1001
        title:
          type: string
          example: 김천 김밥축제
        serviceRegionCode:
          $ref: '#/components/schemas/ServiceRegionCode'
        locationText:
          type: string
          example: 경상북도 김천시
        address:
          type: string
          nullable: true
          example: 경상북도 김천시 삼락동 488-1
        latitude:
          type: number
          format: double
          nullable: true
          example: 36.1398
        longitude:
          type: number
          format: double
          nullable: true
          example: 128.1136
        operatingHours:
          type: string
          nullable: true
        operatingPeriod:
          type: string
          nullable: true
        closedDays:
          type: string
          nullable: true
        usageFee:
          type: string
          nullable: true
        parkingInfo:
          type: string
          nullable: true
        images:
          type: array
          maxItems: 4
          description: |
            상세 화면 갤러리 이미지입니다. 운영자 승인 사진공모전 수상작, 기존 대표사진,
            KTO detailImage2 순으로 우선순위를 적용하고 같은 URL은 한 번만 반환합니다.
            일반 장소는 사용 가능한 만큼만 반환할 수 있지만, 공개된 온보딩 후보 장소는
            고유 이미지 4장이 확인된 경우에만 발행되므로 항상 order 1~4의 정확히 4장을 반환합니다.
          items:
            $ref: '#/components/schemas/PlaceImage'
        tags:
          type: array
          items:
            type: string
        isSaved:
          type: boolean
          description: 익명 조회에서는 항상 false
        description:
          allOf:
            - $ref: '#/components/schemas/PlaceDescription'
          nullable: true
          description: 검증된 원문 설명이 없으면 null
        relatedPlaces:
          type: array
          items:
            $ref: '#/components/schemas/RelatedPlace'
        availableTabs:
          type: array
          items:
            $ref: '#/components/schemas/AvailablePlaceTab'
...
    AvailablePlaceTab:
      type: string
      description: 장소 상세에서 실제로 노출 가능한 탭
      enum: [DESCRIPTION, ROUTE, MATES]
...
    LocationSummary:
      type: object
      required: [locationId, displayName, serviceRegionCode]
      properties:
        locationId:
          type: integer
          format: int64
          example: 100
        displayName:
          type: string
          example: 성신여자대학교
        serviceRegionCode:
          $ref: '#/components/schemas/ServiceRegionCode'
...
    PlaceCard:
      type: object
      required: [placeId, title, serviceRegionCode, serviceRegionName, addressSummary, tags, saved]
      properties:
        placeId:
          type: integer
          format: int64
          example: 1001
        title:
          type: string
          example: 김천 김밥축제
        serviceRegionCode:
          $ref: '#/components/schemas/ServiceRegionCode'
        serviceRegionName:
          type: string
          example: 경상
        addressSummary:
          type: string
          example: 경상북도 김천시
        imageUrl:
          type: string
          format: uri
          nullable: true
          description: null이면 프론트 기본 이미지를 사용
          example: https://cdn.koready.kr/places/1001/cover.jpg
        festivalOccurrence:
          allOf:
            - $ref: '#/components/schemas/FestivalOccurrenceSummary'
          nullable: true
          description: 축제 카드에서만 존재합니다. 개최 회차의 연도·기간·현재 상태를 한 묶음으로 사용합니다.
        travelStyle:
          type: string
          allOf:
            - $ref: '#/components/schemas/TravelStyle'
          nullable: true
        tags:
          type: array
          description: 검수된 태그가 없으면 빈 배열
          items:
            type: string
          example: [지역축제, 로컬음식, 시즌추천]
        shortDescription:
          type: string
          nullable: true
          example: 김밥을 주제로 먹고, 만들고, 즐기는 지역 축제예요.
        saved:
          type: boolean
          description: 현재 로그인 사용자의 저장 여부. 익명 조회에서는 항상 false
...
    FestivalOccurrenceSummary:
      type: object
      description: 반복 축제의 한 개최 회차. status는 저장값이 아니라 Asia/Seoul 조회일 기준으로 계산합니다.
      required: [occurrenceId, eventYear, startDate, endDate, status, dateRangeText]
      properties:
        occurrenceId:
          type: integer
          format: int64
          description: 연도별 개최 회차를 식별하는 내부 ID
        eventYear:
          type: integer
          minimum: 2000
          maximum: 2100
          description: 개최 시작일 기준 연도
        startDate:
          type: string
          format: date
        endDate:
          type: string
          format: date
        status:
          $ref: '#/components/schemas/FestivalOccurrenceStatus'
        dateRangeText:
          type: string
          description: 한국어는 M.d - M.d, 영어는 MMM d, yyyy - MMM d, yyyy 형식인 표시 문자열. 상태 계산에는 사용하지 않습니다.
          example: 7.25 - 7.26
...
    RecommendationDeckEnvelope:
      allOf:
        - $ref: '#/components/schemas/ApiEnvelope'
        - type: object
          properties:
            data:
              $ref: '#/components/schemas/RecommendationDeckResponse'
    RecommendationDeckResponse:
      type: object
      required: [deckId, scope, originLocation, cards, hasMore, remainingThreshold, deduplication]
      properties:
        deckId:
          type: string
          description: 이벤트 전송과 다음 페이지 조회에 사용하는 opaque ID
          example: rec_3f2504e0-4f89-41d3-9a0c-0305e82c3301
        scope:
          $ref: '#/components/schemas/RecommendationScope'
        originLocation:
          $ref: '#/components/schemas/LocationSummary'
        cards:
          type: array
          description: 서버 추천 순서 그대로 렌더링하며 프론트에서 shuffle하지 않음
          items:
            $ref: '#/components/schemas/RecommendationCard'
        nextCursor:
          type: string
          nullable: true
          example: reccur_7d444840-9dc0-11d1-b245-5ffdce74fad2
        hasMore:
          type: boolean
        remainingThreshold:
          type: integer
          minimum: 1
          description: 남은 카드가 이 수 이하가 되면 다음 페이지를 미리 요청
          example: 5
        deduplication:
          $ref: '#/components/schemas/RecommendationDeduplication'
    RecommendationCard:
      type: object
      required: [placeId, title, locationText, saved, tags, serviceRegionCode, matchRank, matchReason]
      properties:
        placeId:
          type: integer
          format: int64
          example: 1001
        title:
          type: string
          example: 김천 김밥축제
        locationText:
          type: string
          example: 경상북도 김천시
        imageUrl:
          type: string
          format: uri
          nullable: true
        saved:
          type: boolean
        tags:
          type: array
          items:
            type: string
          example: [지역축제, 로컬음식, 체험]
        shortDescription:
          type: string
          nullable: true
        serviceRegionCode:
          $ref: '#/components/schemas/ServiceRegionCode'
        travelStyle:
          type: string
          allOf:
            - $ref: '#/components/schemas/TravelStyle'
          nullable: true
        matchRank:
          type: integer
          enum: [1, 2, 3]
          description: 1=태그+유형 일치, 2=유형만 일치, 3=탐색 확장 후보
        matchReason:
          $ref: '#/components/schemas/RecommendationMatchReason'
    RecommendationMatchReason:
      type: object
      required: [travelStyleMatched, preferenceTagMatched, matchedTagCodes]
      properties:
        travelStyleMatched:
          type: boolean
        preferenceTagMatched:
          type: boolean
        matchedTagCodes:
          type: array
          items:
            type: string
          example: [LOCAL_FOOD, HANDS_ON]
    RecommendationDeduplication:
      type: object
      required: [guaranteedWithinDeck, suppressionDays]
      properties:
        guaranteedWithinDeck:
          type: boolean
          enum: [true]
        suppressionDays:
          type: integer
          enum: [30]
          description: CARD_SERVED 뒤 같은 장소의 재노출 금지 기간
    RecommendationEventEnvelope:
      allOf:
        - $ref: '#/components/schemas/ApiEnvelope'
        - type: object
          properties:
            data:
              $ref: '#/components/schemas/RecommendationEventResponse'
    RecommendationEventResponse:
      type: object
      required: [eventId, deckId, placeId, eventType, recordedAt]
      properties:
        eventId:
          type: string
          example: recevt_01J2ABCDEF
        deckId:
          type: string
        placeId:
          type: integer
          format: int64
        eventType:
          $ref: '#/components/schemas/RecommendationEventType'
        recordedAt:
          type: string
          format: date-time
    SavedPlaceListEnvelope:
      allOf:
        - $ref: '#/components/schemas/ApiEnvelope'
        - type: object
          properties:
            data:
              $ref: '#/components/schemas/SavedPlaceListResponse'
    SavedPlaceListResponse:
      type: object
      required: [items, hasMore]
      properties:
        items:
          type: array
          items:
            $ref: '#/components/schemas/SavedPlaceCard'
        nextCursor:
          type: string
          nullable: true
        hasMore:
          type: boolean
    SavedPlaceCard:
      allOf:
        - $ref: '#/components/schemas/PlaceCard'
        - type: object
          required: [savedAt]
          properties:
            savedAt:
              type: string
              format: date-time
              description: 저장 목록 정렬 기준
    SavePlaceEnvelope:
      allOf:
        - $ref: '#/components/schemas/ApiEnvelope'
        - type: object
          properties:
            data:
              $ref: '#/components/schemas/SavePlaceResponse'
    SavePlaceResponse:
      type: object
      required: [placeId, saved, savedAt]
      properties:
        placeId:
          type: integer
          format: int64
        saved:
          type: boolean
          enum: [true]
        savedAt:
          type: string
          format: date-time
    PlaceImage:
      type: object
      required: [imageUrl, order]
      properties:
        imageUrl:
          type: string
          format: uri
        order:
          type: integer
          minimum: 1
        altText:
          type: string
          nullable: true
    PlaceDescription:
      type: object
      required: [impactTitle, impactSubtitle, introParagraphs, enjoyPoints, sourceType]
      properties:
        impactTitle:
          type: string
          example: 김천에서 만나는
        impactSubtitle:
          type: string
          example: 가장 맛있는 한 줄 여행
        introParagraphs:
          type: array
          items:
            type: string
          example: [김밥을 주제로 먹고 만들고 즐길 수 있는 지역 축제예요.]
        enjoyPoints:
          type: array
          items:
            type: string
          example: [다양한 김밥 맛보기, 김밥 만들기 체험, 포토존에서 사진 찍기]
        sourceType:
          type: string
          enum: [KTO_ORIGINAL, AI_GENERATED, MANUAL_EDITED]
          description: 화면에는 노출하지 않아도 되는 콘텐츠 계보 값
    RelatedPlace:
      type: object
      required: [placeId, title, shortDescription]
      properties:
        placeId:
          type: integer
          format: int64
        title:
          type: string
          example: 직지사
        imageUrl:
          type: string
          format: uri
          nullable: true
        shortDescription:
          type: string
          example: 김천에서 함께 둘러보기 좋은 전통 명소예요.
...
    ApiFieldError:
      type: object
      required: [field, reason]
      properties:
        field:
          type: string
          description: 요청 body 또는 query의 필드 경로
          example: travelStyles
        rejectedValue:
          type: string
          nullable: true
          description: 보안상 안전한 경우에만 반환하는 거절 값
        reason:
          type: string
          example: 1개 이상 4개 이하로 선택해야 합니다.
...